simplified package structuring
