--- clibase/main.go
@@ -1,7 +1,0 @@
-package main
-
-import "github.com/iter8-tools/iter8/cmd"
-
-func main() {
-	cmd.Execute()
-}
