--- clibase/main_test.go
@@ -1,9 +1,0 @@
-package main
-
-import (
-	"testing"
-)
-
-func TestMain(t *testing.T) {
-	main()
-}
