--- k8s/cmd/root.go
@@ -1,18 +1,0 @@
-package cmd
-
-import (
-	basecli "github.com/iter8-tools/iter8/cmd"
-)
-
-var RootCmd = basecli.RootCmd
-
-// Execute adds all child commands to the root command and sets flags appropriately.
-// This is called by main.main(). It only needs to happen once to the rootCmd.
-func Execute() {
-	RootCmd.Execute()
-}
-
-func init() {
-	// extend base gen command with the k8s command
-	basecli.GenCmd.AddCommand(NewGetK8sCmd())
-}
