+++ main.go
@@ -1,6 +1,6 @@
 package main
 
-import "github.com/iter8-tools/iter8/k8s/cmd"
+import "github.com/iter8-tools/iter8/cmd"
 
 func main() {
 	cmd.Execute()
